+++ changelog.md
@@ -3,3 +3,4 @@
 # 2018-08-11
 - Added minimum type of characters functionality.
 - Fixed bug to prevent user from selecting more minimum characters then password length.
+- Fixed bug to prevent user from de selecting letters and requiring first character to be a letter.